requirements.txt
aiogram==3.13.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
aiofiles==24.1.0
httpx==0.28.1